requirements.txt
fastapi
uvicorn
pandas
matplotlib
groq
google-generativeai
python-dotenv
streamlit
requests
